Eagerly add commands to protocol stack #2456
To avoid promise completion reordering causing a different protocol stack order, we now eagerly add commands to the protocol stack,
